Remove root component, move initialization into index, add index for actions
