--- components/Root.js
@@ -1,32 +1,0 @@
-import React, { Component } from 'react';
-import { createStore, applyMiddleware } from 'redux';
-import { Provider } from 'react-redux';
-import { composeWithDevTools } from 'redux-devtools-extension/developmentOnly';
-import thunkMiddleware from 'redux-thunk';
-import rootReducer from 'reducers';
-import App from 'components/App';
-
-const composeEnhancers = composeWithDevTools({
-    name: 'ReduxBoilerplate',
-});
-
-const store = createStore(
-    rootReducer,
-    composeEnhancers(
-        applyMiddleware(
-            thunkMiddleware
-        )
-    )
-);
-
-class Root extends Component {
-    render() {
-        return (
-            <Provider store={store}>
-                <App />
-            </Provider>
-        );
-    }
-};
-
-export default Root;
