+++ data/actions/index.js
@@ -1,0 +1,5 @@
+import * as actionTypes from './types';
+
+export {
+    actionTypes,
+};
