+++ data/reducers/index.js
@@ -1,5 +1,7 @@
 import { combineReducers } from 'redux';
 
-const rootReducer = combineReducers({});
+const rootReducer = combineReducers({
+
+});
 
 export default rootReducer;
